Backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
requests==2.31.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.3
h5py>=3.10.0
numpy>=1.26.0
filelock>=3.13.0
